+++ lib/sisimai/smtp/reply.rb
@@ -1,4 +1,5 @@
 # http://www.ietf.org/rfc/rfc5321.txt
+# -------------------------------------------------------------------------------------------------
 #   4.2.1.  Reply Code Severities and Theory
 #
 #   There are four values for the first digit of the reply code:
@@ -103,8 +104,9 @@ module Sisimai
           # 554   Transaction failed (Or, in the case of a connection-opening response, "No SMTP service here")
           # 555   MAIL FROM/RCPT TO parameters not recognized or not implemented
           # 556   Domain does not accept mail (See RFC7504)
+          # 557   draft-moore-email-addrquery-01
           '550', '552', '553', '551', '521', '525', '502', '520', '523', '524', '530', '533', '534',
-          '535', '538', '551', '555', '556', '554', '500', '501', '502', '503', '504',
+          '535', '538', '551', '555', '556', '554', '557', '500', '501', '502', '503', '504',
         ].freeze
         CodeOfSMTP = { '2' => ReplyCode2, '4' => ReplyCode4, '5' => ReplyCode5 }.freeze
 
@@ -119,8 +121,8 @@ module Sisimai
           reply = argv1.to_i
           first = (reply / 100).to_i
 
-          return false if reply < 200
-          return false if reply > 599
+          return false if reply < 211
+          return false if reply > 557
           return false if reply % 100 > 59
 
           if first == 2